KS-21-04-write-to-file 1) Перенос кода из файля для задачи 2, в файл для этой задачи
